+++ src/osha/oira/upgrade/v1/20230228134447_add_newsletter_statistics/upgrade.py
@@ -1,0 +1,9 @@
+from ftw.upgrade import UpgradeStep
+from osha.oira.upgrade.utils import alembic_upgrade_statistics_to
+
+
+class AddNewsletterStatistics(UpgradeStep):
+    """Add newsletter statistics."""
+
+    def __call__(self):
+        alembic_upgrade_statistics_to(self.target_version)